Single lane stoplights implemented. Each lane gets its own stoplight (can theoretically work for multi-lane roads, but for now designed with just single-lane roads in mind) and a car can get the color of the light at the end of its current edge.

+++ core/src/com/mygdx/sim/GameObjects/data/Node.java
@@ -10,7 +10,7 @@ public class Node implements Comparable<Node> {
 	private Coordinates location;
 	private double nodeDistanceWeight;
 	private double nodeDistanceWeightEstimate;
-	private int nodePriorityWeight;
+	private double nodePriorityWeight;
 	private boolean isDestination;
 
 	public Node(double xCoordinate, double yCoordinate) {
@@ -69,7 +69,7 @@ public class Node implements Comparable<Node> {
 		return previousNode;
 	}
 
-	public int getNodePriorityWeight() {
+	public double getNodePriorityWeight() {
 		return nodePriorityWeight;
 	}
